pyproject.toml
[project]
name = "unet-from-scratch-in-pytorch"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "numpy>=2.3.3",
    "pillow>=11.3.0",
    "tqdm>=4.67.1",
]
